Enhance Order and Product classes with new methods
Added a factory method `Create` to the `Order` class to facilitate the creation of `Order` objects with specified properties and to raise an `OrderCreatedEvent`.
Introduced an `Update` method in the `Order` class to update the order details and raise an `OrderUpdatedEvent`.
Added methods `Add` and `Remove` to the `Order` class to manage `OrderItem` objects within an order.
Modified the constructor of the `OrderItem` class to change the order of parameters, placing `orderId` after `productId`.
Added a factory method `Create` to the `Product` class to facilitate the creation of `Product` objects with specified properties.